Mega PROJECT 1. JARVIS/requirements.txt
SpeechRecognition==3.10.1
pyttsx3==2.90
PyAudio==0.2.14
wikipedia==1.4.0
pyjokes==0.6.0
psutil==5.9.8
pywhatkit==5.4
pyautogui
speedtest-cli
requests
